Create EnvironmentAlteration
An idea to affect change on the environment using time-based scoring. I plan to code it so you place an object, like an Oxygen Vent, and it'll give you Oxy points over time. We want to have it so OxyScore won't rise when the object isn't on a cell, multiple Oxygen Vents will compound, and maybe some other stuff.